refactor: move NEOs state and useEffect to NEOGrid
