+++ components/NEOGrid.tsx
@@ -1,15 +1,39 @@
-import React, { useRef, useEffect } from 'react';
+import React, { useRef, useEffect, useState } from 'react';
 import { StyleSheet, View, ScrollView } from 'react-native';
 import { NEO as NEOType } from '../types';
 import { NEO } from './NEO';
+import { getNEOsByDate } from '../services/NASA/NEOWebService';
 
 interface NEOGridProps {
-  NEOs: NEOType[];
+  selectedDate: Date
 }
 
-export const NEOGrid = ({ NEOs }: NEOGridProps) => {
+export const NEOGrid = ({ selectedDate }: NEOGridProps) => {
+  const [NEOs, setNEOs] = useState<NEOType[]>([]);
   const scrollRef = useRef<ScrollView>(null);
   const getRandomPosition = () => Math.floor(Math.random() * 3); // 0, 1, or 2
+
+  // Fetch NEOs for the selected date when date changes
+  useEffect(() => {
+    const fetchNEOs = async () => {
+      const date = selectedDate.toISOString().split('T')[0];
+      const NEOSResponse = await getNEOsByDate(date);
+      const NEOSResult = NEOSResponse.map((NEO: any) => {
+
+        const diameter = (NEO.estimated_diameter.feet.estimated_diameter_min + NEO.estimated_diameter.feet.estimated_diameter_max) / 2;
+        
+        return {
+          name: NEO.name,
+          diameter: diameter,
+          velocity: NEO.close_approach_data[0].relative_velocity.miles_per_hour,
+          distance: NEO.close_approach_data[0].miss_distance.miles,
+          isHazardous: NEO.is_potentially_hazardous_asteroid
+        }
+      })
+      setNEOs(NEOSResult);
+    }
+    fetchNEOs();
+  }, [selectedDate]);
 
   // Reset scroll to the top when NEOs change
   useEffect(() => {
